separate endpoints, prompts and create enums
